Making corrections from comments
