+++ drf/accounts/urls.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
 from django.urls import include, path
 from rest_framework_jwt.views import obtain_jwt_token, refresh_jwt_token
-from accounts import views
 from .views import LoginView
 
 urlpatterns = [
