+++ drf/drf/settings.py
@@ -154,7 +154,7 @@ JWT_AUTH = {
 
     'JWT_RESPONSE_PAYLOAD_HANDLER':
     # 'rest_framework_jwt.utils.jwt_response_payload_handler',
-    'accounts.api.utils.jwt_response_payload_handler',
+    'accounts.utils.jwt_response_payload_handler',
 
     'JWT_ALLOW_REFRESH': True,
     'JWT_REFRESH_EXPIRATION_DELTA': datetime.timedelta(days=7),
